Improve performance by 250%, add better shaders
The Vertex Shader now computes distance to the player's light source.
The method that was used before broke when I batched all layers together
(see more info below).

---

The performance on my i7-9700k + RTX 2070 in windowed mode was absolutely
despicable (~25 FPS).

This was because each Tile's layer was drawn with one draw call each.
All layers of a tile are now batched into one call.

Here's a rough CPU performance breakdown of the debug build:
The following values are all in percent of total time taken in a ~3 minute run
of the game in debug mode with a profiler attached.
I dismissed all times regarding initialization, but the worst offender in
initialization is `ResourceManager::make_procedural_texture`.

* `Renderer::render` uses 80-90% of our time, since we dont have
barely any logic yet.

    * `ResourceManager::get_full_shader` - 2.5%
    * `ResourceManager::get_procedural_texture` - 3.3%
    * Iteration of all tiles during rendering and other operations - 9%
    * `sf::RenderTarget::draw` - 69% (nice.)
        * `sf::Shader::bind` - two different calls, one 18.5%, the other 12%

* `CollisionEngine::update`, in case anyone is wondering, takes up about 0.4%
of the frame's time (i.e. nothing).
    * `handle_collision` - 0.2%